And ... move some more to ui
