+++ src/app/GUI/canvaswindow.cpp
@@ -32,7 +32,7 @@
 #include "mainwindow.h"
 #include "GUI/BoxesList/boxscroller.h"
 #include "swt_abstraction.h"
-#include "renderoutputwidget.h"
+#include "dialogs/renderoutputwidget.h"
 #include "Sound/soundcomposition.h"
 #include "GUI/global.h"
 #include "renderinstancesettings.h"
