+++ src/ui/widgets/animationwidgetscrollbar.cpp
